fix: ci action version audit, docs corrections, and pre-commit guard
- unify actions/checkout to v6 and upload-artifact to v7 across all workflows
- add swatinem/rust-cache and concurrency control to fuzz.yml
- add rust-cache to release.yml post-release-check
- add loc check to pre-commit hook (max 500 lines)
- source cargo env in pre-commit hook for reliable toolchain access
- extract sample-app tests to src/tests.rs to pass loc gate
- fix docs: dora-metrics workflow ref, context.yaml placeholders, ci.md tiers
- update readme: version badge to 0.3.0, crate list, feature defaults, benchmarks
- add verify-actions skill with evals for ci action sha verification
