+++ .github/workflows/dora-fdrt.yml
@@ -16,7 +16,7 @@ jobs:
       issues: read
     steps:
       - name: Checkout repository
-        uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5  # v4
+        uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd  # v6
 
       - name: Calculate FDRT
         uses: actions/github-script@f28e40c7f34bde8b3046d885e986cb6290c5673b  # v7
@@ -53,7 +53,7 @@ jobs:
             fs.writeFileSync('dora-metrics.jsonl', JSON.stringify(metric) + '\n');
 
       - name: Upload DORA Metrics
-        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02  # v4
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a  # v7
         with:
           name: dora-metrics-fdrt-${{ github.event.issue.number }}
           path: dora-metrics.jsonl
