+++ .github/workflows/hotfix.yml
@@ -55,7 +55,7 @@ jobs:
             fs.writeFileSync('dora-metrics.jsonl', JSON.stringify(event) + '\n');
 
       - name: Upload DORA Metrics
-        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02  # v4
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a  # v7
         with:
           name: dora-metrics-hotfix-${{ github.run_id }}
           path: dora-metrics.jsonl
